CSP-882 Moved network events page to sharedui
