+++ src/SFA.DAS.Aan.SharedUi.UnitTests/Extensions/DateTimeExtensionsTests.cs
@@ -1,4 +1,5 @@
-﻿using FluentAssertions;
+﻿using AutoFixture.NUnit3;
+using FluentAssertions;
 using SFA.DAS.Aan.SharedUi.Extensions;
 
 namespace SFA.DAS.Aan.SharedUi.UnitTests.Extensions;
@@ -10,7 +11,25 @@ public class DateTimeExtensionsTests
     public void UtcToLocalTime(DateTimeKind kind)
     {
         var date = new DateTime(2023, 5, 7, 13, 10, 0, kind);
-        var actual = date.SharedUtcToLocalTime();
+        var actual = date.UtcToLocalTime();
         actual.Hour.Should().Be(14);
+    }
+
+    [Test, AutoData]
+    public void ToApiString_ReturnsFormattedDateOnlyString(DateTime date)
+    {
+        DateTimeExtensions.ToApiString(DateOnly.FromDateTime(date)).Should().Be(date.ToString("yyyy-MM-dd"));
+    }
+
+    [Test, AutoData]
+    public void ToApiString_ReturnsFormattedDateTimeString(DateTime date)
+    {
+        date.ToApiString().Should().Be(date.ToString("yyyy-MM-dd"));
+    }
+
+    [Test, AutoData]
+    public void ToScreenString_ReturnsFormattedDateTimeString(DateTime date)
+    {
+        date.ToScreenString().Should().Be(date.ToString("dd/MM/yyyy"));
     }
 }
