+++ src/SFA.DAS.Aan.SharedUi.UnitTests/UrlHelpers/MailtoLinkValueTests.cs
@@ -1,17 +1,18 @@
 ﻿using SFA.DAS.Aan.SharedUi.UrlHelpers;
+
 namespace SFA.DAS.Aan.SharedUi.UnitTests.UrlHelpers;
 
-public class MailtoLinkValueTests
+public class MailToLinkValueTests
 {
     [Test]
-    public void FromAddressAndSubject_CreatesMailtoLink()
+    public void FromAddressAndSubject_CreatesMailToLink()
     {
         const string emailAddress = "me@email.com";
         const string subject = "This Is Very Important";
         const string expected = $"mailto:{emailAddress}?subject=This%20Is%20Very%20Important";
 
-        var mailtoLink = MailtoLinkValue.FromAddressAndSubject(emailAddress, subject);
+        var mailToLink = MailToLinkValue.FromAddressAndSubject(emailAddress, subject);
 
-        Assert.That(mailtoLink, Is.EqualTo(expected));
+        Assert.That(mailToLink, Is.EqualTo(expected));
     }
 }
