+++ src/SFA.DAS.Aan.SharedUi/Models/NetworkEvents/FilterFields.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.ApprenticeAan.Web.Models.NetworkEvents;
+﻿namespace SFA.DAS.Aan.SharedUi.Models.NetworkEvents;
 
 public enum FilterFields
 {
