+++ src/SFA.DAS.Aan.SharedUi/Services/QueryStringParameterBuilder.cs
@@ -1,5 +1,5 @@
-﻿using SFA.DAS.ApprenticeAan.Web.Extensions;
-using SFA.DAS.ApprenticeAan.Web.Models.NetworkEvents;
+﻿using SFA.DAS.Aan.SharedUi.Extensions;
+using SFA.DAS.Aan.SharedUi.Models.NetworkEvents;
 
 namespace SFA.DAS.ApprenticeAan.Web.Services;
