+++ src/SFA.DAS.Aan.SharedUi/UrlHelpers/MailtoLinkValue.cs
@@ -1,6 +1,6 @@
 ﻿namespace SFA.DAS.Aan.SharedUi.UrlHelpers;
 
-public static class MailtoLinkValue
+public static class MailToLinkValue
 {
     public static string FromAddressAndSubject(string emailAddress, string subject)
     {
