+++ src/SFA.DAS.Aan.SharedUi/Views/NetworkEvents/_CalendarEvents.cshtml
@@ -1,8 +1,7 @@
 ﻿﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
-@using SFA.DAS.ApprenticeAan.Domain.Constants
-@using SFA.DAS.ApprenticeAan.Web.Extensions;
-@using SFA.DAS.Aan.SharedUi.Constants;
+@using SFA.DAS.Aan.SharedUi.Models.NetworkEvents;
 @using SFA.DAS.Aan.SharedUi.Extensions;
+@using SFA.DAS.Aan.SharedUi.Constants;
 
 @model NetworkEventsViewModel
 <ol class="das-search-results__list">
