+++ src/SFA.DAS.Aan.SharedUi/Views/NetworkEvents/_FilterDetails.cshtml
@@ -1,5 +1,6 @@
-﻿@using SFA.DAS.ApprenticeAan.Web.Models.NetworkEvents
-@model SFA.DAS.ApprenticeAan.Web.Models.NetworkEvents.ChecklistDetails
+﻿@using SFA.DAS.Aan.SharedUi.Models.NetworkEvents;
+
+@model ChecklistDetails
 @{
     var counter = 0;
 }
