+++ src/SFA.DAS.Aan.SharedUi/Views/NetworkEvents/_FilterEventDates.cshtml
@@ -1,4 +1,7 @@
-﻿@model SFA.DAS.ApprenticeAan.Web.Models.NetworkEvents.EventFilterChoices
+﻿@using SFA.DAS.Aan.SharedUi.Extensions;
+
+@model SFA.DAS.ApprenticeAan.Web.Models.NetworkEvents.EventFilterChoices
+
 <div class="govuk-form-group">
     <fieldset class="govuk-fieldset">
         <legend class="govuk-fieldset__legend govuk-fieldset__legend--m">
