+++ src/SFA.DAS.ApprenticeAan.Web.UnitTests/SFA.DAS.ApprenticeAan.Web.UnitTests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
@@ -38,6 +38,8 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\SFA.DAS.Aan.SharedTestHelpers\SFA.DAS.Aan.UnitTests.Shared.csproj" />
+    <ProjectReference Include="..\SFA.DAS.Aan.SharedUi.UnitTests\SFA.DAS.Aan.SharedUi.UnitTests.csproj" />
     <ProjectReference Include="..\SFA.DAS.ApprenticeAan.Web\SFA.DAS.ApprenticeAan.Web.csproj" />
   </ItemGroup>
