+++ src/SFA.DAS.ApprenticeAan.Web.UnitTests/Services/FilterBuilderTests.cs
@@ -4,8 +4,9 @@ using Microsoft.AspNetCore.Mvc.Routing;
 using Moq;
 using SFA.DAS.Aan.SharedUi.Constants;
 using SFA.DAS.Aan.SharedUi.Extensions;
+using SFA.DAS.Aan.SharedUi.Models.NetworkEvents;
+using SFA.DAS.Aan.SharedUi.Services;
 using SFA.DAS.ApprenticeAan.Web.Models.NetworkEvents;
-using SFA.DAS.ApprenticeAan.Web.Services;
 
 namespace SFA.DAS.ApprenticeAan.Web.UnitTests.Services;
