+++ src/SFA.DAS.ApprenticeAan.Web.sln
@@ -21,7 +21,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.ApprenticeAan.Domai
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Aan.SharedUi", "SFA.DAS.Aan.SharedUi\SFA.DAS.Aan.SharedUi.csproj", "{E5FDA0D3-D7E6-4B06-B6A4-F7899CA77A01}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SFA.DAS.Aan.SharedUi.UnitTests", "SFA.DAS.Aan.SharedUi.UnitTests\SFA.DAS.Aan.SharedUi.UnitTests.csproj", "{E021C0CB-9E7A-4059-9134-98A507F16CC4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Aan.SharedUi.UnitTests", "SFA.DAS.Aan.SharedUi.UnitTests\SFA.DAS.Aan.SharedUi.UnitTests.csproj", "{E021C0CB-9E7A-4059-9134-98A507F16CC4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SFA.DAS.Aan.UnitTests.Shared", "SFA.DAS.Aan.SharedTestHelpers\SFA.DAS.Aan.UnitTests.Shared.csproj", "{4C2BB4E5-048E-4977-96B6-399F1C554130}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -65,6 +67,10 @@ Global
 		{E021C0CB-9E7A-4059-9134-98A507F16CC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E021C0CB-9E7A-4059-9134-98A507F16CC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E021C0CB-9E7A-4059-9134-98A507F16CC4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4C2BB4E5-048E-4977-96B6-399F1C554130}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4C2BB4E5-048E-4977-96B6-399F1C554130}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4C2BB4E5-048E-4977-96B6-399F1C554130}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4C2BB4E5-048E-4977-96B6-399F1C554130}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -75,6 +81,7 @@ Global
 		{6C335DA2-0E25-4795-BFFB-F35B2F037BD4} = {60CC78EB-71AE-462B-BFC8-8DE07C9D4C96}
 		{7E69F529-1F65-4CB1-BDF2-36F234D842CB} = {60CC78EB-71AE-462B-BFC8-8DE07C9D4C96}
 		{E021C0CB-9E7A-4059-9134-98A507F16CC4} = {60CC78EB-71AE-462B-BFC8-8DE07C9D4C96}
+		{4C2BB4E5-048E-4977-96B6-399F1C554130} = {60CC78EB-71AE-462B-BFC8-8DE07C9D4C96}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4A73F9BF-B39B-4624-A3E9-2DE9DEE32CC5}
